First attempt at refactoring
There's still a lot I would like to change.
However, there's a lot of code to take in.
I think it would be best to change things gradually to prevent problems
I've added a lot of comments throughout some of the most glaring issues.
I also started work on moving the code into smaller files.

This bot will currently not function on this branch.